+++ test/test.js
@@ -16,6 +16,7 @@ describe('i18next', function() {
       lowerCaseLng: false,
       ns: 'translation',
       fallbackToDefaultNS: false,
+      fallbackOnNull: true,
       resGetPath: 'locales/__lng__/__ns__.json',
       dynamicLoad: false,
       useLocalStorage: false,
@@ -762,7 +763,7 @@ describe('i18next', function() {
       };
       
       beforeEach(function(done) {
-        i18n.init(i18n.functions.extend(opts, { resStore: resStore, returnObjectTrees: true }),
+        i18n.init(i18n.functions.extend(opts, { resStore: resStore, returnObjectTrees: true, fallbackOnNull: false }),
           function(t) { done(); });
       });
     
@@ -770,6 +771,27 @@ describe('i18next', function() {
         expect(i18n.t('key1')).to.be(null);
         expect(i18n.t('key2')).to.eql({ key3: null });
       });
+    
+      describe('with option fallbackOnNull = true', function() {
+        var resStore = {
+          dev: { translation: { key1: 'fallbackKey1', key2: { key3: 'fallbackKey3' } } },
+          en: { translation: { } },            
+          'en-US': { translation: { key1: null, key2: { key3: null } } }
+        };
+        
+        beforeEach(function(done) {
+          i18n.init(i18n.functions.extend(opts, { resStore: resStore, fallbackOnNull: true}),
+            function(t) { done(); });
+        });
+    
+        it('it should translate to fallback value', function() {
+          expect(i18n.t('key1')).to.be('fallbackKey1');
+          expect(i18n.t('key2.key3')).to.eql('fallbackKey3');
+        });
+    
+      });
+    
+    
     });
   
     describe('key with empty string value as valid option', function() {